requirements.txt
django==2.0.8
djangorestframework==3.9.0
graphene-django==2.0
